plugins/armnn/src/requirements.txt
tflite-runtime==2.14.0
pillow==10.3.0